+++ src/components/content/content.tsx
@@ -1,13 +1,37 @@
-import React from "react";
+import React, {useEffect} from "react";
 import styled from "styled-components";
 import {Calendar} from "@local/components/calendar";
 import {BoomWidget} from "@local/components/widget";
 import {breakpoints} from "@local/components/theme/breakpoints";
 import {LocaleSwitcher} from "@local/components/layout/locale-switcher";
+import {useIframeMessenger} from "@local/hooks/iframe-messenger";
+import {useAppContext} from "@local/runtime";
 
 export function Content(): JSX.Element {
+    const { organizerId } = useAppContext();
+    const { handleContentInitialized, handleContentResize  } = useIframeMessenger();
+    const contentId = 'grid-content';
+
+    const handleResize = (_entries: ResizeObserverEntry[]) => {
+        handleContentResize(organizerId!);
+    }
+
+    /** Run after page init necessarily steps to handle widget on 3rd party website */
+    const handleWidgetMessageBroker = () => {
+        handleContentInitialized(organizerId!);
+        const element = document.getElementById(contentId);
+        if (element) {
+            new ResizeObserver(handleResize).observe(element);
+        } else {
+            console.warn("handleWidgetMessageBroker - element not found!");
+        }
+    }
+
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+    useEffect(handleWidgetMessageBroker, []);
+
     return (
-        <Grid>
+        <Grid id={contentId}>
             <Left>
                 <LocaleSwitcher />
                 <Calendar />
